# [ITensors] [BUG] Fix for performance issues when applying gates to every other site