--- a/src/mps/abstractmps.jl
+++ b/src/mps/abstractmps.jl
@@ -2087,11 +2087,15 @@ expτH = ops(os, s)
 ```
 """
 function product(
-  As::Vector{<:ITensor}, ψ::AbstractMPS; move_sites_back::Bool=true, kwargs...
+  As::Vector{<:ITensor},
+  ψ::AbstractMPS;
+  move_sites_back_between_gates::Bool=false,
+  move_sites_back::Bool=true,
+  kwargs...
 )
   Aψ = ψ
   for A in As
-    Aψ = product(A, Aψ; move_sites_back=false, kwargs...)
+    Aψ = product(A, Aψ; move_sites_back=move_sites_back_between_gates, kwargs...)
   end
   if move_sites_back
     s = siteinds(Aψ)