@mtfishman 2022-04-15
@GTorlai we will have to think about the implications of this for `runcircuit` in PastaQ.